added readme for AStar, better structure for examples
